requirements.txt
aioredis==1.3.1
amqp==5.0.9
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.3.1
async-timeout==4.0.2
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
Brotli==1.0.9
cached-property==1.5.2
ce==0.0.0.0.dev20200924
celery==5.0.5
certifi==2020.6.20
cffi==1.14.6
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
colorlog==6.5.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
daphne==3.0.2
dataclasses==0.8
decorator==5.1.1
Deprecated==1.2.13
Django==2.2.16
django-celery==3.3.1
django-celery-beat==2.2.0
django-celery-results==2.2.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-filter==2.4.0
django-ranged-response==0.2.0
django-redis==4.12.1
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.4.1
django-restframework==0.0.1
django-simple-captcha==0.5.13
django-timezone-field==4.2.3
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
dnspython==1.16.0
drf-extensions==0.7.0
drf-yasg==1.20.0
enum34==1.1.10
eventlet==0.30.2
Flask==1.1.4
flower==1.0.0
future==0.18.2
gevent==21.1.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
h2==4.0.0
hiredis==1.1.0
hpack==4.0.0
humanize==3.14.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.8.3
incremental==21.3.0
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.3
jsonpath==0.82
kaitaistruct==0.9
kombu==5.1.0
ldap3==2.8.1
lxml==4.6.5
Markdown==3.3.6
MarkupSafe==0.23
mitmproxy==5.3.0
mongoengine==0.22.1
monotonic==1.6
more-itertools==8.12.0
msgpack==1.0.3
msgpack-python==0.4.8
numpy==1.19.5
packaging==21.3
paramiko==2.9.2
passlib==1.7.4
Pillow==8.1.0
pluggy==0.13.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.13.0
psutil==5.8.0
publicsuffix2==2.20191221
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
Pygments==2.7.4
PyJWT==1.7.1
pymongo==3.11.3
PyMySQL==0.10.0
PyNaCl==1.5.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.2
pytest==5.4.3
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.24.0
retry==0.9.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.15.0
sortedcontainers==2.2.2
SQLAlchemy==1.4.31
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
Twisted==21.7.0
txaio==21.2.1
typing==3.7.4.3
typing_extensions==4.1.1
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.25.11
urwid==2.1.2
user-agents==2.2.0
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.13.3
wsproto==0.15.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
zstandard==0.14.1

watchdog~=2.1.7
